pyproject.toml
[project]
name = "fast-crud"
version = "0.1.0"
description = "Base FastAPI router with CRUD operations, based on SQLModel entities"
readme = "README.md"
dependencies = [
    "aiohttp>=3.11.14",
    "alembic>=1.15.1",
    "cryptography>=44.0.2",
    "faker>=37.1.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "orjson>=3.10.16",
    "psycopg[binary]>=3.2.6",
    "pydantic-settings[yaml]>=2.8.1",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "pynacl>=1.5.0",
    "pytest-asyncio>=0.26.0",
    "pytest-freezegun>=0.4.2",
    "pytest>=8.3.5",
    "pytz>=2025.2",
    "redis>=5.2.1",
    "sqids>=0.5.1",
    "sqlmodel>=0.0.24",
    "uuid6>=2024.7.10",
]